Trello Card: https://trello.com/c/IFwoA3pt/539-refactoring-core-bfft-mymain
Cleaned code based on reviews provided.
