+++ BFFT_Architecture_STM32_FreeRTOS.ioc
@@ -280,7 +280,7 @@ VP_FREERTOS_VS_CMSIS_V2.Signal=FREERTOS_VS_CMSIS_V2
 Mcu.Pin28=PG13
 Mcu.IP6=USART3
 Mcu.Pin29=PB3
-ProjectManager.CoupleFile=false
+ProjectManager.CoupleFile=true
 RCC.48MHZClocksFreq_Value=24000000
 PB3.Signal=SYS_JTDO-SWO
 RCC.SYSCLKFreq_VALUE=144000000
@@ -307,7 +307,6 @@ Mcu.Pin15=PD9
 NVIC.HardFault_IRQn=true\:0\:0\:false\:false\:true\:false\:true\:false
 Mcu.Pin16=PG6
 Mcu.Pin13=PB14
-PD0.GPIO_PuPd=GPIO_PULLUP
 Mcu.Pin14=PD8
 Mcu.Pin19=PA9
 NVIC.TIM1_UP_TIM10_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:true
@@ -322,7 +321,6 @@ RCC.PLLQ=6
 NVIC.PriorityGroup=NVIC_PRIORITYGROUP_4
 Mcu.Pin11=PB0
 Mcu.Pin12=PB13
-PD0.GPIOParameters=GPIO_PuPd
 RCC.PLLM=4
 RCC.PLLN=144
 Mcu.Pin10=PC5
