+++ Core/Src/stm32f7xx_it.c
@@ -186,10 +186,12 @@ void CAN1_RX0_IRQHandler(void)
   /* USER CODE BEGIN CAN1_RX0_IRQn 1 */
   //CAN1_irq_receive();
   CAN_RxHeaderTypeDef header;
-  char data[8];
-  if (HAL_CAN_GetRxMessage(&hcan1, 0, &header, data) != HAL_OK)
+  uint8_t u8Data[8];
+  if (HAL_CAN_GetRxMessage(&hcan1, 0, &header, u8Data) != HAL_OK)
   {
-
+     asm("NOP");
+     asm("NOP");
+     asm("NOP");
   }
   /* USER CODE END CAN1_RX0_IRQn 1 */
 }
