+++ Debug/Core/BFFT/testCAN.d
@@ -1,5 +1,5 @@
-Core/BFFT/middleware_layer/can/CANController.o: \
- ../Core/BFFT/middleware_layer/can/CANController.cpp \
+Core/BFFT/testCAN.o: ../Core/BFFT/testCAN.cpp \
+ E:/SoftwareWorkingFolder/BlackForestFormulaTeam/branches/Refactoring_VCU_Core_BFFT_myMain/Core/BFFT/device_layer/inverter/Inverter.hpp \
  E:/SoftwareWorkingFolder/BlackForestFormulaTeam/branches/Refactoring_VCU_Core_BFFT_myMain/Core/BFFT/middleware_layer/can/CANController.hpp \
  E:/SoftwareWorkingFolder/BlackForestFormulaTeam/branches/Refactoring_VCU_Core_BFFT_myMain/Core/BFFT/middleware_layer/can/CANControllerBase.hpp \
  E:/SoftwareWorkingFolder/BlackForestFormulaTeam/branches/Refactoring_VCU_Core_BFFT_myMain/Core/BFFT/middleware_layer/can/ICANController.hpp \
@@ -38,7 +38,14 @@ Core/BFFT/middleware_layer/can/CANController.o: \
  ../Drivers/STM32F7xx_HAL_Driver/Inc/stm32f7xx_hal_tim.h \
  ../Drivers/STM32F7xx_HAL_Driver/Inc/stm32f7xx_hal_tim_ex.h \
  ../Drivers/STM32F7xx_HAL_Driver/Inc/stm32f7xx_hal_uart.h \
- ../Drivers/STM32F7xx_HAL_Driver/Inc/stm32f7xx_hal_uart_ex.h
+ ../Drivers/STM32F7xx_HAL_Driver/Inc/stm32f7xx_hal_uart_ex.h \
+ ../Core/BFFT/device_layer/inverter/InverterLeft.hpp \
+ ../Core/BFFT/device_layer/inverter/Inverter.hpp \
+ ../Core/BFFT/device_layer/inverter/InverterRight.hpp \
+ E:/SoftwareWorkingFolder/BlackForestFormulaTeam/branches/Refactoring_VCU_Core_BFFT_myMain/Core/BFFT/testCAN.hpp \
+ ../Core/Inc/main.h
+
+E:/SoftwareWorkingFolder/BlackForestFormulaTeam/branches/Refactoring_VCU_Core_BFFT_myMain/Core/BFFT/device_layer/inverter/Inverter.hpp:
 
 E:/SoftwareWorkingFolder/BlackForestFormulaTeam/branches/Refactoring_VCU_Core_BFFT_myMain/Core/BFFT/middleware_layer/can/CANController.hpp:
 
@@ -117,3 +124,13 @@ E:/SoftwareWorkingFolder/BlackForestFormulaTeam/branches/Refactoring_VCU_Core_BF
 ../Drivers/STM32F7xx_HAL_Driver/Inc/stm32f7xx_hal_uart.h:
 
 ../Drivers/STM32F7xx_HAL_Driver/Inc/stm32f7xx_hal_uart_ex.h:
+
+../Core/BFFT/device_layer/inverter/InverterLeft.hpp:
+
+../Core/BFFT/device_layer/inverter/Inverter.hpp:
+
+../Core/BFFT/device_layer/inverter/InverterRight.hpp:
+
+E:/SoftwareWorkingFolder/BlackForestFormulaTeam/branches/Refactoring_VCU_Core_BFFT_myMain/Core/BFFT/testCAN.hpp:
+
+../Core/Inc/main.h:
