+++ Debug/Core/Src/can.su
@@ -1,0 +1,3 @@
+can.c:30:6:MX_CAN1_Init	8	static
+can.c:62:6:HAL_CAN_MspInit	88	static
+can.c:113:6:HAL_CAN_MspDeInit	16	static
