+++ Debug/Core/Src/gpio.su
@@ -1,0 +1,1 @@
+gpio.c:55:6:MX_GPIO_Init	56	static
