+++ Debug/Core/Src/stm32f7xx_hal_msp.su
@@ -1,5 +1,1 @@
 stm32f7xx_hal_msp.c:64:6:HAL_MspInit	16	static
-stm32f7xx_hal_msp.c:88:6:HAL_CAN_MspInit	48	static
-stm32f7xx_hal_msp.c:140:6:HAL_CAN_MspDeInit	16	static
-stm32f7xx_hal_msp.c:174:6:HAL_UART_MspInit	48	static
-stm32f7xx_hal_msp.c:210:6:HAL_UART_MspDeInit	16	static
