+++ Debug/Core/Src/stm32f7xx_it.su
@@ -5,7 +5,7 @@ stm32f7xx_it.c:116:6:BusFault_Handler	4	static
 stm32f7xx_it.c:131:6:UsageFault_Handler	4	static
 stm32f7xx_it.c:146:6:DebugMon_Handler	4	static
 stm32f7xx_it.c:166:6:CAN1_TX_IRQHandler	8	static
-stm32f7xx_it.c:180:6:CAN1_RX0_IRQHandler	48	static
-stm32f7xx_it.c:200:6:CAN1_RX1_IRQHandler	8	static
-stm32f7xx_it.c:214:6:CAN1_SCE_IRQHandler	8	static
-stm32f7xx_it.c:228:6:TIM1_UP_TIM10_IRQHandler	8	static
+stm32f7xx_it.c:180:6:CAN1_RX0_IRQHandler	48	static,ignoring_inline_asm
+stm32f7xx_it.c:202:6:CAN1_RX1_IRQHandler	8	static
+stm32f7xx_it.c:216:6:CAN1_SCE_IRQHandler	8	static
+stm32f7xx_it.c:230:6:TIM1_UP_TIM10_IRQHandler	8	static
