+++ Debug/Core/Src/usart.su
@@ -1,0 +1,3 @@
+usart.c:31:6:MX_USART3_UART_Init	8	static
+usart.c:61:6:HAL_UART_MspInit	48	static
+usart.c:91:6:HAL_UART_MspDeInit	16	static
