+++ Debug/objects.list
@@ -1,18 +1,22 @@
 "Core/BFFT/control_system_layer/double_buffering/test.o"
 "Core/BFFT/device_layer/inverter/Inverter.o"
+"Core/BFFT/device_layer/inverter/InverterLeft.o"
+"Core/BFFT/device_layer/inverter/InverterRight.o"
 "Core/BFFT/frutil/Ticks.o"
 "Core/BFFT/frutil/error.o"
 "Core/BFFT/frutil/mutex_errors.o"
 "Core/BFFT/frutil/test.o"
 "Core/BFFT/frutil/test2.o"
 "Core/BFFT/heap_useNewlib.o"
-"Core/BFFT/test.o"
+"Core/BFFT/testCAN.o"
 "Core/BFFT/middleware_layer/can/CANController.o"
 "Core/BFFT/middleware_layer/can/CANFrame.o"
 "Core/BFFT/middleware_layer/can/CANFrameId.o"
 "Core/BFFT/middleware_layer/can/CANPayload.o"
 "Core/BFFT/middleware_layer/can/src/test.o"
+"Core/Src/can.o"
 "Core/Src/freertos.o"
+"Core/Src/gpio.o"
 "Core/Src/main.o"
 "Core/Src/stm32f7xx_hal_msp.o"
 "Core/Src/stm32f7xx_hal_timebase_tim.o"
@@ -20,6 +24,7 @@
 "Core/Src/syscalls.o"
 "Core/Src/sysmem.o"
 "Core/Src/system_stm32f7xx.o"
+"Core/Src/usart.o"
 "Core/Startup/startup_stm32f767zitx.o"
 "Drivers/STM32F7xx_HAL_Driver/Src/stm32f7xx_hal.o"
 "Drivers/STM32F7xx_HAL_Driver/Src/stm32f7xx_hal_can.o"
